requirements.txt
certifi==2024.7.4
chardet==4.0.0
coloredlogs==15.0
humanfriendly==9.1
idna==3.7
jsonschema==3.2.0
packaging==21.3
pexpect==4.8.0
ptyprocess==0.7.0
pyparsing==3.0.9
PyYAML==6.0.1
requests==2.32.4
toposort==1.6
urllib3==2.5.0
pytest==7.4.4
python-debian==1.0.1